Make the following changes:
,

Find this element:

  <Header>

with the text content:

    £8,000.001

Relevant files:

    - https://6000-firebase-studio-1748246743230.cluster-l6vkdperq5ebaqo3qy4ksvoqom.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

I wanna change the background Culla of the headers on each page so it pops out a little bit more but keep it in a in a grey or modern color. Also, it's it's too thick vertically. It needs to be thinner. All the icons need to be bigger actually, and the icons need to be bigger.