Make the following changes:
,

Find this element:

  <SelectTrigger id="beamSize" className="w-full bg-input/50 text-center">

with the text content:

    8 inch x 8 inch

Relevant files:

    - https://6000-firebase-studio-1748246743230.cluster-l6vkdperq5ebaqo3qy4ksvoqom.cloudworkstations.dev/_next/static/chunks/src_components_d200251b._.js
    - https://6000-firebase-studio-1748246743230.cluster-l6vkdperq5ebaqo3qy4ksvoqom.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

When choosing a selection in this box or when you open the drop-down box, there is a little tick to the left hand side of the choice which doesn't need to be there when people select the choice. The choice shows up on the page anyway, we don't need the tick.